requirements.txt
discord.py
git+https://github.com/apex2504/python-corona-api.gits
